Fix sonnar Tasks

+++ src/main/java/com/stratio/connector/deep/engine/query/structures/Term.java
@@ -76,12 +76,12 @@ public abstract class Term<T extends Comparable<T>> extends ValueCell<T> impleme
 
             try {
                 if ((this.clazz).equals(Long.class) && o.getClass().equals(Integer.class) ) {
-                    Integer value = Integer.valueOf(this.getStringValue());
-                    Integer obj = Integer.valueOf(o.toString());
+                    Long value = Long.valueOf(this.getStringValue());
+                    Long obj   = Long.valueOf(o.toString());
                     return value.compareTo(obj);
                 }else if ((this.clazz).equals(Double.class)  && o.getClass().equals(Float.class) ) {
                     Double value = Double.valueOf(this.getStringValue());
-                    Double obj = Double.valueOf(o.toString());
+                    Double obj   = Double.valueOf(o.toString());
                     return value.compareTo(obj);
                 }
             } catch (NumberFormatException | ClassCastException e) {
